% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inla.R
\name{ind.fit}
\alias{ind.fit}
\title{Metrics from INLA}
\usage{
ind.fit(inlaObj)
}
\arguments{
\item{inlaObj}{object returned by \code{inla}}
}
\value{
list
}
\description{
Return useful metriucs from INLA (DIC, WAIC).
}
\author{
Marie Denis
}

